requirements.txt
typer[all]
ord-schema
rdkit-pypi
pandas
beautifulsoup4
lxml
requests
sqlalchemy
